extract module and fix types
