+++ protostar/cli/protostar_argument.py
@@ -1,4 +1,5 @@
-from .argument import Argument
+from protostar.argument_parser import Argument
+
 from .protostar_arg_type import ProtostarArgTypeName
 
 ProtostarArgument = Argument[ProtostarArgTypeName]
