+++ protostar/cli/protostar_command.py
@@ -1,4 +1,5 @@
-from .command import Command
+from protostar.argument_parser import Command
+
 from .protostar_arg_type import ProtostarArgTypeName
 
 ProtostarCommand = Command[ProtostarArgTypeName]
