+++ protostar/composition_root.py
@@ -3,8 +3,8 @@ from logging import getLogger
 from pathlib import Path
 from typing import Optional
 
-from protostar.cli import ArgumentParserFacade, map_protostar_type_name_to_parser
-from protostar.cli.protostar_command import ProtostarCommand
+from protostar.argument_parser import ArgumentParserFacade
+from protostar.cli import ProtostarCommand, map_protostar_type_name_to_parser
 from protostar.commands import (
     BuildCommand,
     DeclareCommand,
