+++ protostar/start.py
@@ -3,7 +3,11 @@ import sys
 from pathlib import Path
 from typing import Any, Optional
 
-from protostar.cli import ArgumentParserFacade, CLIApp, MissingRequiredArgumentException
+from protostar.argument_parser import (
+    ArgumentParserFacade,
+    CLIApp,
+    MissingRequiredArgumentException,
+)
 from protostar.composition_root import build_di_container
 from protostar.configuration_profile_cli import ConfigurationProfileCLI
 from protostar.protostar_cli import ProtostarCLI
